Use thread local cnt & bytes
CSPPMemTab::m_wals[*].cnt & bytes is not accurate if do not use
atomic fetch_add, there are ~2% on db_bench fillrandom with 8 threads.

If we use atomic, it should be slow, so we use thread local counter
and flush to global in batch(THREAD_LOCAL_THRESHOLD).

On ConvertToReadOnly, flush all tls_token.